requirements.txt
streamlit==1.35.0
google-genai>=1.0.0
psycopg2-binary>=2.9.0
pyngrok==7.2.0
